+++ .yarnrc
@@ -2,4 +2,5 @@
 # yarn lockfile v1
 
 
-yarn-path ".yarn/releases/yarn-1.18.0.cjs"
+lastUpdateCheck 1669515161763
+yarn-path ".yarn/releases/yarn-1.16.0.js"
