+++ backend/package.json
@@ -26,7 +26,9 @@
     "firebase": "^8.0.1",
     "firebase-admin": "^9.3.0",
     "fuse.js": "^6.4.6",
+    "ip": "^1.1.8",
     "kafka-node": "^5.0.0",
+    "kafkajs": "^2.2.3",
     "mongoose": "^6.6.7",
     "mongoose-autopopulate": "^0.16.1",
     "morgan": "^1.10.0",
