--- common/constants/index.ts
@@ -1,3 +1,0 @@
-export const REVIEW_CHARACTER_LIMIT = 2000;
-
-export const REVIEW_PHOTOS_LIMIT = 2000;
