+++ docker-compose.yml
@@ -1,0 +1,30 @@
+version: '3'
+
+services:
+  backend:
+    env_file: './backend/.env'
+    container_name: back_container
+    build:
+      ./backend
+      # context: ./backend
+      # dockerfile: ./Dockerfile
+
+    image: 'ithaca-apts/backend'
+    ports:
+      - '8080:8080'
+
+  frontend:
+    env_file: './frontend/.env'
+    container_name: front_container
+    build:
+      ./frontend
+      # context: ./frontend
+      # dockerfile: ./Dockerfile
+    image: 'ithaca-apts/frontend'
+    environment:
+      PORT: 3000
+      # PROXY_API: http://playground-web-backend:8080/
+    ports:
+      - '3000:3000'
+    links:
+      - 'backend'
