+++ frontend/src/components/Apartment/Header.tsx
@@ -9,7 +9,7 @@ import {
   Avatar,
 } from '@material-ui/core';
 import styles from './Header.module.scss';
-import { ApartmentWithId } from '../../../../common/types/db-types';
+import { ApartmentWithId } from '../../../common/types/db-types';
 import defaultHeader from '../../assets/default_header.png';
 import defaultIcon from '../../assets/default_icon.png';
 import { colors } from '../../colors';
