+++ frontend/src/components/Home/Autocomplete.tsx
@@ -11,7 +11,7 @@ import {
   Link,
 } from '@material-ui/core';
 import { get } from '../../utils/call';
-import { LandlordOrApartmentWithLabel } from '../../../../common/types/db-types';
+import { LandlordOrApartmentWithLabel } from '../../../common/types/db-types';
 import SearchIcon from '@material-ui/icons/Search';
 import { makeStyles } from '@material-ui/core/styles';
 import { Link as RouterLink } from 'react-router-dom';
