+++ frontend/src/components/Landlord/Header.tsx
@@ -10,7 +10,7 @@ import {
   Avatar,
 } from '@material-ui/core';
 import styles from './Header.module.scss';
-import { Landlord } from '../../../../common/types/db-types';
+import { Landlord } from '../../../common/types/db-types';
 import defaultHeader from '../../assets/default_header.png';
 import defaultIcon from '../../assets/default_icon.png';
 import { colors } from '../../colors';
