+++ frontend/src/pages/LandlordPage.tsx
@@ -10,11 +10,11 @@ import { useTitle } from '../utils';
 import LandlordHeader from '../components/Landlord/Header';
 import { get } from '../utils/call';
 import styles from './LandlordPage.module.scss';
-import { Landlord } from '../../../common/types/db-types';
+import { Landlord } from '../../common/types/db-types';
 import Toast from '../components/LeaveReview/Toast';
 import LinearProgress from '../components/utils/LinearProgress';
 // import { Likes, ReviewWithId } from '../../../common/types/db-types';
-import { ReviewWithId } from '../../../common/types/db-types';
+import { ReviewWithId } from '../../common/types/db-types';
 import { createAuthHeaders, subscribeLikes, getUser } from '../utils/firebase';
 import DropDown from '../components/utils/DropDown';
 import NotFoundPage from './NotFoundPage';
